+++ HISTORY.rst
@@ -7,6 +7,7 @@ Pending release
 ---------------
 
 * New release notes here
+* Reset the random state for NumPy too.
 
 1.2.3 (2016-08-19)
 ------------------